chore: moved utils
